changed spacing on stat page
