+++ EZWorkout.xcworkspace/xcuserdata/spencersymington.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -58,7 +58,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "EZWorkout/stat VC/ScrollStatsView.swift"
-            timestampString = "574031769.0138839"
+            timestampString = "574038303.948609"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "10"
